Move mapreduce files from main
